% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model.R
\name{add_noise_func}
\alias{add_noise_func}
\title{Adds observation error}
\usage{
add_noise_func(x, params)
}
\arguments{
\item{x}{the value to be perturbed}

\item{params}{vector of parameters for the observation error matrix. params[1] should be the value for S, and params[2] should be the value for EA. Params[3] should be the maximum observable titre.}
}
\value{
returns the given value with added observation error
}
\description{
Adds random noise to a given value, based on the shouldered observation error distribution.
}
\examples{
add_noise_func(5, c(0.79,0.2,15))
}

