% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analyses.R
\name{calculate_statistics}
\alias{calculate_statistics}
\title{Calculate infection statistics from MCMC chains}
\usage{
calculate_statistics(chain_names, all_ind_pars, all_data, burnin,
  threshold = 0.5)
}
\description{
Make sure you're in the working directory with all of the chains
}

