% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model.R
\name{convert_data_list_to_table}
\alias{convert_data_list_to_table}
\title{Converts list to table}
\usage{
convert_data_list_to_table(all_data, all_ind_pars, strainNames = c("A", "B",
  "C"))
}
\arguments{
\item{all_data}{list of data frames for each individual}

\item{all_ind_pars}{list of parameter lists}

\item{strainNames}{vector of strain names}
}
\value{
a data frame of formatted data
}
\description{
Given a list as returned from simulate_and_save, converts to the data format of fluscape
}

