% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model.R
\name{convert_data_table_to_list}
\alias{convert_data_table_to_list}
\title{Convert data to list}
\usage{
convert_data_table_to_list(dat, nstrains = 3, n_samples = 2,
  strainNames = c("A", "B", "C"))
}
\arguments{
\item{dat}{data frame of the data to be converted}

\item{nstrains}{Number of strains}

\item{n_samples}{number of time samples}

\item{strainNames}{vector of strain names}
}
\value{
a list of the data
}
\description{
Given a table of data (like fluscape), converts this into a list for use in MCMC
}

