% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model.R
\name{cum_inc}
\alias{cum_inc}
\title{Cumulative incidence conversion}
\usage{
cum_inc(inc)
}
\arguments{
\item{inc}{a vector of incidence probability}
}
\value{
cumulative incidence vector
}
\description{
Given an incidence vector, gives the cumulative probability of infection over the same period
}

