% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model.R
\name{fluscape_simulation}
\alias{fluscape_simulation}
\title{FluScape serology simulation}
\usage{
fluscape_simulation(fluscape_data, n = NULL, fluscape_strains, mu_pars,
  tp_pars, m_pars, STOCHASTIC = TRUE, incidenceVectors,
  removeOutliers = TRUE, fluscapeT0 = TRUE, addNoise = TRUE,
  plotSerology = FALSE, noiseParams = c(0.79, 0.2, 8))
}
\arguments{
\item{fluscape_data}{the FluScape data frame, as returned by load.and.merge.part.V1.V2()}

\item{n}{Number of individuals to simulate. If NULL, uses the entire population}

\item{fluscape_strains}{a vector of influenza strains to be used. Note that this should just be the strain ID. For example, "H3N2.2008". The function will create identifiers to access the V1 and V2 values.}

\item{mu_pars}{n x n matrix of boosting parameters, where n is the number of simulated strains. Each strain should provide some boost (may be 0) to each other strain.}

\item{tp_pars}{n x n matrix of time to peak parameters}

\item{m_pars}{n x n matrix of waning parameters. Waning rate may be set to zero.}

\item{STOCHASTIC}{boolean value indicating whether or not boosting should be stochastic or deterministic. If stochastic, each boost is drawn from a poisson distribution.}

\item{incidenceVectors}{a list of incidence vectors. Should match length of fluscape_strains (ie. each circulating strain has an incidence vector. Note that the range of these vectors should cover the FluScape study period. ie. go from 0 (the day of the first sample) to x (the day of the last sample).}

\item{removeOutliers}{boolean value indicating whether or not sample time outliers from the fluscape data should be included. Defaults to TRUE}

\item{fluscapeT0}{boolean value indicating whether or not the V1 titres should be used as t0 titres in the simulation. If not, all startnig titres are set to 0. Defaults to TRUE}

\item{addNoise}{boolean value indicating whether or not observation error noise should be added to the simulation. Defaults to TRUE}

\item{plotSerology}{boolean value indicating whether or not the FluScape data and simulation data should be plotted. Defaults to FALSE.}

\item{noiseParams}{vector of parameters to be passed to \code{\link{add_noise}}}
}
\value{
a dataframe of a similar format to the FluScape data frame.
}
\description{
Simulates HAI titre readings to match the FluScape serosurvey protocol using the simplified boosting and waning model.
}
\examples{
fluscape_dat <- load.and.merge.V1.V2()

mu_pars <- matrix(ncol=2,nrow=2)
mu_pars[1,] <- c(5,1)
mu_pars[2,] <- c(1,6)

tp_pars <- matrix(ncol=2,nrow=2)
tp_pars[1,] <- tp_pars[,2] <- c(21,21)

m_pars <- matrix(ncol=2,nrow=2)
m_pars[1,] <- m_pars[2,] <- c(0,0)

inc_vectors <- NULL
inc_vectors[[1]] <- generate_toy_inc(0.4,900,200,600)
inc_vectors[[2]] <- generate_toy_inc(0.1,900,500,750)
 
fluscape_simulation(fluscape_dat, 100, c("H3N2.2008","CKH9N2.2008"), mu_pars, tp_pars, m_pars, TRUE, inc_vectors, TRUE, TRUE, FALSE, TRUE, c(0.8,0.1))
Get given strains into format for fluscape data frame
Get only data that we need - sampling times and titres
Omit NA
Starting titres
}
\seealso{
\code{\link{overall_simulation}}
}

