% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model.R
\name{generate_infection_time}
\alias{generate_infection_time}
\title{Random infection time sample}
\usage{
generate_infection_time(cuminc)
}
\arguments{
\item{cuminc}{cumulative incidence vector}
}
\value{
a single infection time
}
\description{
Generates a random infection time from a given incidence vector
}

