% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model.R
\name{generate_toy_inc}
\alias{generate_toy_inc}
\title{Toy step incidence generator}
\usage{
generate_toy_inc(p_inf, end, t0, t1)
}
\arguments{
\item{p_inf}{Proportion of populatino that should become infected}

\item{end}{final day of the vector (ie. length)}

\item{t0}{first day of possible infection}

\item{t1}{last day of possible infection}
}
\value{
vector with probability of infection on each day
}
\description{
Given an probability of infection, generates a uniform distribution giving probability of infection on each day. From day 0 to t0, has 0 probability of infection. t0 to t1 has finite probability. t1 to end has 0 probability.
}
\examples{
generate_toy_inc(0.4, 200, 50, 150)
}
\seealso{
\code{\link{generate_toy_inc}}
}

