% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model.R
\name{generate_uniform_inc}
\alias{generate_uniform_inc}
\title{Uniform incidence generator}
\usage{
generate_uniform_inc(p_inf, t1, t2)
}
\arguments{
\item{p_inf}{Proportion of population that should be infected}

\item{t1}{starting day}

\item{t2}{end day}
}
\value{
vector with probability of infection on each day
}
\description{
Given an overall proportion of the population that should become infected, this function generates a uniform probability of infection between two time points.
}
\examples{
generate_uniform_inc(0.4, 1,100)
}
\seealso{
\code{\link{generate_toy_inc}}
}

