% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model.R
\name{generateIncCurve}
\alias{generateIncCurve}
\title{SIR inc curve}
\usage{
generateIncCurve(t, startPops, params)
}
\description{
Generates an incidence curve from an SIR model. Same parameters as \code{\link{SIRsim}}.
}

