% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analyses.R
\name{get_infection_ratios}
\alias{get_infection_ratios}
\title{Get posterior propn}
\usage{
get_infection_ratios(chain, burnin, threshold, tmin, tmax)
}
\description{
Calculates proportion of posterior estimate from chain that indicates infection
}

