% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model.R
\name{individual_simulation}
\alias{individual_simulation}
\title{Individual serology simulation}
\usage{
individual_simulation(infection_times, y0s, processParams, times,
  strain_names = c("a", "b", "c"), sample_times,
  PROCESS_FUNCTION = multiple_strains)
}
\arguments{
\item{infection_times}{a vector of infection times for each strain (index should match strain names order)}

\item{processParams}{a generic vector of parameters to be passed to the data generating function.}

\item{times}{a vector of times to return values from the simulation}

\item{strain_names}{a vector of strain names for data labels}

\item{PROCESS_FUNCTION}{pointer to a function to calculate the data at the given time points. The function should take the following arguments: "infection_times" - a vector of times of infection with each strain; "y0" - a vector of initial values for each strain; processParams - as above; t - vector of times. This function should return a matrix of serological values matching the simulation data type. The first column should be time in days, and each subsequent column should be the measurement for each strain of interest on that day. See \code{\link{multiple_strains}} for an example (and default).}

\item{y0}{vector of baseline values. eg. if three pathogens are being tested, this should be a vector of three values.}
}
\value{
a data frame with time series readings for each of the given strains
}
\description{
Simulates a data frame of serology measurements using the given data generating function.
}
\examples{

infection_times <- c(0,50,100)
y0 <- c(0,0,0)

times <- seq(0,200,by=1)
strains <- c("A","B","C")

dat <- individual_simulation(infection_times, y0, processParams, times, strains, multiple_strains)
Go through each infection and generate a trajectory
Create column names and only return desired sample times
}
\seealso{
\code{\link{overall_simulation}}
}

