% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model.R, R/old.R
\name{multiple_strains}
\alias{multiple_strains}
\title{Wrapper for the model}
\usage{
multiple_strains(tis, y0s, mu_pars, tp_pars, m_pars, times)

multiple_strains(tis, y0s, mu_pars, tp_pars, m_pars, times)
}
\arguments{
\item{tis}{vector of times of infection eg. c(0, 45, 70)}

\item{y0s}{vector of starting titre values}

\item{mu_pars}{matrix of boosting parameters}

\item{tp_pars}{matrix of time to peak parameters}

\item{m_pars}{matrix of waning parameters}

\item{times}{vector of times in days.}

\item{infection_times}{vector of integer infection times}

\item{processParams}{list of parameters for the model}

\item{t}{vector of time points to be solved over}

\item{y0s}{vector of baseline HAI values for each infecting strain}
}
\value{
a matrix of the titre values for each time and corresponding times

matrix of times and longitundinal simulation for each infecting strain
}
\description{
Wrapper function for the Rcpp implementation of the multi strain boosting/waning model

Data generating function for a multiple strain, boosting and waning serological model.
}
\seealso{
\code{\link{single_strain}}
}

