% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model.R
\name{overall_simulation}
\alias{overall_simulation}
\title{Serosurvey simulation function}
\usage{
overall_simulation(individuals = 100, strains = c("H3N2", "H5N1", "H1N1"),
  strainIncidences, samplingTimes, processParams, y0s, startTime = 0,
  endTime = 200, addNoise = TRUE, noiseParams = c(0.79, 0.2, 8),
  discreteData = TRUE, PROCESS_FUNCTION = multiple_strains,
  NOISE_FUNCTION = add_noise_func, MIXED_EFFECTS = TRUE)
}
\arguments{
\item{individuals}{number of individuals to simulate. Defaults to 100}

\item{strains}{a vector of virus strain names to label the simulation data}

\item{strainIncidences}{a list of incidence vectors. Should match length of the strains vector (ie. each circulating strain has an incidence vector. Note that the range of these vectors should cover the entire study period. ie. go from 0 (the day of the first sample) to x (the day of the last sample).}

\item{processParams}{a generic vector of parameters to be passed to the data generating function. If using mixed effects, should also provide the distributions/parameters of the population parameters}

\item{y0s}{baseline readings for each individual. May simply be a matrix of zeroes.}

\item{startTime}{the start time of the survey. Defaults to 0}

\item{endTime}{the end time of the survey. Defaults to 200 (ie. 200 days)}

\item{addNoise}{if true, adds noise to the simulated data}

\item{noiseParams}{generic vector of parameters to pass to the noise generating function. Defaults to match \code{\link{add_noise}}}

\item{discreteData}{if TRUE, takes the floor of all generated data. Defaults to TRUE.}

\item{NOISE_FUNCTION}{pointer to a noise adding function. Should take the value to be manipulated and a vector of function specific parameters (noiseParams). See \code{\link{add_noise}} for an example.}

\item{MIXED_EFFECTS}{bool to indicate whether or not mixed effects should be used}

\item{logTitre}{if TRUE, returns log transformed titre values. (5*2^y). This might be specific to HAI titre data, so might be best left to default as FALSE.}
}
\value{
a dataframe of the simulated serosurvey.
}
\description{
Simulates readings from a generic serosurvey. The function can take pointers to any data generating function and noise adding function to generate values generated from a custom process. A number of bools and parameters can be passed to customise the protocol.
}
\examples{
n <- 1000

inc_vectors <- NULL
inc_vectors[[1]] <- generate_toy_inc(0.4,900,200,600)
inc_vectors[[2]] <- generate_toy_inc(0.1,900,500,750)

samplingTimes <- matrix(ncol=2, nrow=n)
samplingTimes[,1] <- 0
samplingTimes[,2] <- 150

y0s <- matrix(ncol=2,nrow=n)
y0s[,1] <- 0
y0s[,2] <- 0

processParams <- list(mu_pars, tp_pars, m_pars)

y <- overall_simulation(n, c("A","B"), inc_vectors, samplingTimes, processParams, y0s, 0, 200, TRUE, c(0.8,0.1),TRUE,TRUE,multiple_strains, add_noise)
ERROR CHECKS
Create the matrix to store the data
Row for each individual
One column for each time point for each strain; one column for each time point.
Generate a cumulative risk of infection vector for each strain    
For each strain, generate a cumulative incidence vector
For each individual, generate an infection time for each strain and simulate
Generate an infection time for each strain. Make sure that no two infections can occur within 21 days of each other
If data is discrete, take floor of matrix
Add noise and log transform if required
First column in times
}
\seealso{
\code{\link{individual_simulation}}
}

