% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package.R
\docType{package}
\name{ serosim2 }
\alias{ serosim2 }
\alias{ serosim2 }
\title{ Infection time and antibody dynamics model }
\description{
Set of functions implementing an infection time and antibody kinetics model. Enables infection times and antibody dynamics to be inferred using serological data.
}
