% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{plot_serology}
\alias{plot_serology}
\title{Serological data plot}
\usage{
plot_serology(passed_dat, strain, saveImage = FALSE)
}
\arguments{
\item{passed_dat}{Data should be a 4 column data frame: First column time of first sample; second column is values from first sample; third column is time of second sample; fourth column is values from second sample}

\item{strain}{string of strain name for plot (will be used as plot title)}
}
\value{
Returns the total ggplot object as a gridExtra object
}
\description{
Plots of provided  paired serological data in a nice way, including a histogram for both visits; a boxplot for both visits and a scatterplot of V1 vs. V2.
}

