% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model.R
\name{plotSIR}
\alias{plotSIR}
\title{SIR plot}
\usage{
plotSIR(y)
}
\arguments{
\item{y}{4 column data frame of the time and 3 compartments}
}
\description{
Simple plot of a 4 column SIR run - time, S, I and R values.
}

