% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mcmc.R
\name{run_metropolis_MCMC}
\alias{run_metropolis_MCMC}
\title{Run MCMC code}
\usage{
run_metropolis_MCMC(all_data, times, mcmcPars = c(iterations = 1000, popt =
  0.44, opt_freq = 50, thin = 1, burnin = 100, adaptive_period = 100, save_block
  = 500, tuning = 100, N_pop_start = 50), filename, pop_fixed, ind_fixed,
  all_pop_pars, all_ind_pars, ...)
}
\arguments{
\item{placeholder}{asgfa}
}
\description{
runs the MCMC chain on the fluscape model
}

