% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analyses.R
\name{save_all_data}
\alias{save_all_data}
\title{Save data}
\usage{
save_all_data(data)
}
\arguments{
\item{data}{the list of data frames}
}
\value{
a single table of formatted data
}
\description{
Formats list of data to single table
}

