% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analyses.R
\name{save_ini_pars}
\alias{save_ini_pars}
\title{Save ind pars}
\usage{
save_ini_pars(ind_pars, tmax)
}
\arguments{
\item{ind_pars}{the list of parameters}

\item{tmax}{max time corresponding to not infected}
}
\value{
a table of pars
}
\description{
Formats a list of individual parameters to save as table
}

