% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analyses.R
\name{seroconversion_check}
\alias{seroconversion_check}
\title{Check for seroconversion}
\usage{
seroconversion_check(data, y0s)
}
\arguments{
\item{data}{data for each individual (ie. each individual has a matrix of sampling times and titres)}

\item{y0s}{vector of initial titre values (ie. at t = 0)}
}
\value{
a vector of infected/not infected for each strain
}
\description{
Checks for seroconversion within an individual data frame
}

