% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/old.R
\name{single_strain}
\alias{single_strain}
\title{Single strain boost/wane model}
\usage{
single_strain(mu_pars, tp_pars, m_pars, ti_pars, y0, times)
}
\arguments{
\item{mu_pars}{vector of boosting parameters}

\item{tp_pars}{vector of time to peak parameters (days)}

\item{m_pars}{vector of waning rate parameters}

\item{ti_pars}{vector of infection times}

\item{times}{vector of times to run model over}
}
\value{
vector of HAI values at each day of the simulation
}
\description{
Simulates the time series antibody kinetics for a single strain from arbritary infection events
}
\seealso{
\code{\link{multiple_strains}}
}

