% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model.R
\name{multiple_strains}
\alias{multiple_strains}
\title{Wrapper for the model}
\usage{
multiple_strains(infection_times, y0s, processParams, t)
}
\arguments{
\item{infection_times}{vector of integer infection times}

\item{y0s}{vector of starting titre values}

\item{processParams}{list of parameters for the model}

\item{t}{vector of time points to be solved over}
}
\value{
a matrix of the titre values for each time and corresponding times
}
\description{
Wrapper function for the Rcpp implementation of the multi strain boosting/waning model
}

